requirements.txt
easydict==1.6
joblib==0.10.3
numpy
opencv-python==3.4.2.16
Pillow==4.0.0
tensorflow-gpu==1.2.0
